+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 ### Added
 - Support for Bitcoin Cash network block sychronization.
 - Support for Bitcoin Cash signatures.
@@ -33,3 +33,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - Removed `.stylish-haskell.yaml` files.
 - Removed old `haskoin-node` and `haskoin-wallet` packages from main repository.
+- Removed support for non-strict signatures and related tests.
+- Removed script evaluator and related tests.